// src/types.d.ts
export interface ProductDatatype {
  name: string;
  title: string;
  userid: number;
  description: string;
  category: string;
  tag: string;
  size: string;
  weight: number;
  skuId: string;
  colour: string;
}

export interface MediaData {
  name: string;
  description: string;
  category: string;
  url: string;
  userid: number;
}